Update and rename src/SalesManager.java to Src/SalesManager.java

+++ Src/SalesManager.java
@@ -12,9 +12,9 @@ public class SalesManager {
         long max = -1;
         for (long sale : sales) {
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
 
             if (sale > max) {
                 max = sale;
@@ -22,9 +22,9 @@ public class SalesManager {
         }
         return max;
     }
-    public int min() {
-        int min = 0;
-        for (int sale : sales) {
+    public long min() {
+        long min = 0;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -32,9 +32,9 @@ public class SalesManager {
         return min;
 
     }
-    public int average() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long average() {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
 
         }
